fsmc_nand.c: Fixed data abort inside change_bit()
Since change_bit() requires a (unsigned int *) as second arg,
the correct definition of err_idx[] array declared as
local variable of fsmc_correct_data() is the following:

    u32 err_idx[8];

+++ drivers/mtd/nand/fsmc_nand.c
@@ -158,7 +158,7 @@ static int fsmc_correct_data(struct mtd_info *mtd, u_char *dat,
 		      u_char *read_ecc, u_char *calc_ecc)
 {
 	/* The calculated ecc is actually the correction index in data */
-	u16 err_idx[8];
+	u32 err_idx[8];
 	u64 ecc_data[2];
 	u32 num_err, i;
